apis: ping the PostgreSQL connection
Currently, the PostgreSQL connection is managed by the code listening
for notifications, and in case of incorrect credentials/connection
issues, Chichi enters a mode where it prints a lot of output to stderr
and no longer responds to interrupt signals.

This commit, therefore, adds code that pings, simplifying the scenario
where PostgreSQL credentials are not entered correctly.