assets: rename the directory from 'vendor' to 'node_modules_vendor'
Calling the assets directory 'vendor' was causing it not to be copied
into the Go cache. Therefore, when running
'go get github.com/open2b/chichi/assets' outside the repository, the
'go run' command failed to compile the package as it couldn't find the
'vendor' directory.

This change fixes the issue by renaming the directory from 'vendor' to
'node_modules_vendor'.