doc: use snake_case names in right-side of mappings documentation
Even if not directly specified in the documentation, the right side of
the mappings is what a user might initially find on the right when
importing, thus representing the names of the properties in the data
warehouse. Since we've decided to suggest snake_case as the convention
for schema property names, this commit updates the documentation
accordingly.