apis: check the properties conflicts in the datastore
This commit moves the code that performs conflict checking between
properties inside the datastore, as this check needs to understand the
logic by which properties are represented as columns, and this logic is
precisely in the datastore.

This allows us to remove a usage of the "warehouses.PropertiesToColumns"
function, which will eventually be phased out.

This commit is related to issue #734.