apis/datastore: move pkg `diffschemas` from `warehouses` to `datastore`
The "diffschemas" package is currently within "warehouses". The fact is
that the diff between two schemas must be managed by the datastore
(especially when we modify the diff to return columns instead of
properties), and it is not something that should be accessible to the
warehouses.

Therefore, this commit moves the package.