assets: move connection existence check to fix `Not Found` redirect
This commit addresses an issue introduced by commit cf2e624, where
creating a new connection resulted in a redirect to the "Not Found"
route.

This occurs because at the moment the connection is created and the
redirect to the connection route happens, it is not yet present in the
application state, as it will be available only after a slight delay when
the connections are fetched from the server again.

For this reason, this commit moves the check for the existence of the
connection back where it was before commit cf2e624.