apis/datastore: simplify the `DataWarehouse.Close` method
This change simplifies the 'Close' method of 'DataWarehouse' by ensuring
that when it is called, no ongoing calls to the data warehouse are in
progress and no new calls will be initiated.

It also removes an outdated comment on the 'Records.Err' method and
updates the comment of the '*APIS.Close' method.