apis,doc: restrict acceptance to standard UUID format only
Previously, UUIDs from connectors were accepted in any of the formats
accepted by the 'uuid.Parse' method, namely:

    "xxxx-xxxx-xxxx-xxxxxxxxxxxx"
    "urn:uuid:xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
    "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
    "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"

where '*' represents any byte.

In filters, only the previous formats were accepted, with the last one
limited to "{xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx}".

This change ensures that only the standard format
"xxxx-xxxx-xxxx-xxxxxxxxxxxx" is accepted, and in the case of filters
alone, uppercase letters are not allowed.

For #760