apis: skip exported users with missing properties
This change prevents exported users with missing properties from being
written to the connector's writer, addressing a bug where the 'Create'
and 'Update' methods require at least one property.

Additionally, this change updates the documentation to clarify that the
'Writer.Write' method in the 'connectors' package expects at least one
property to be provided, and that both the '(*Transformer).Transform'
and '(*Mapping).Transform' functions can return an empty map for
properties.

For #769
Closes #658