assets: delay the computation of the grid's columns widths
This commit ensures that the widths of the columns in the grids are
calculated with a slight delay to ensure that all elements within the
cells have been fully rendered.

The brief additional period during which the widths have not yet been
calculated is managed by hiding the grid content to avoid flashes of
unstyled content.