all: store a single Anonymous ID per identity instead of multiple
With the current implementation of the Connection Identity Resolution,
it is impossible for an identity to have multiple anonymous IDs.
Therefore, this commit ensures that only one anonymous ID is stored per
identity.

This change simplifies the code, the UI and documentation. It is
important to note that the code within the Connection Workspace Identity
Resolution will be removed once identity writing through Merge is
implemented; this commit also modifies that code for consistency and to
keep the tests functioning.