chichi: move role from connector config field to method parameters
This change removes the 'role' attribute from direct storage in the
config structs. Instead, it's now passed as a parameter to the 'Schema',
'ServeUI', and 'ReceiveWebhook' methods.

With this change, when a connector instance is created, it doesn't know
its role until one of its methods is called. While most methods assume
the role implicitly, for 'Schema', 'ServeUI', and 'ReceiveWebhook', it
is provided as an argument.

In the 'ReceiveWebhook' method, if the webhooks are for connector or
resource, the role is 'Both'.