apis/events/collector: prevent client interruptions of event processing
Previously, if the client sending events received a response and
subsequently closed the connection, it would result in the interruption
of identity updates.

Now, upon receiving events, the collector ensures uninterrupted
processing regardless of whether the client closes the HTTP connection,
whether due to receiving the response or any other reason.

Note that if the client closes the HTTP connection before receiving the
response, it may attempt to resend the events. However, in such cases,
the events are identified as duplicates and are not processed again.