all: add an ordering property path to actions when exporting to files
This commit adds an ordering property path to actions when exporting to
files, which determines the sorting of exported users.

The change in the UI is implemented separately on the 'main', therefore
at the moment through the UI it is not possible to add actions that
export to files. In this regard, see issue #788.

Resolve #757.