apis, doc: rename the primary key column in table "users_identities"
The column is renamed to make clearer that it has no logical value, but
it is only used at the implementation level.

Theoretically it can be removed to use (1) the connection, (2) the
identity ID and (3) the is/is not anonymous column, but at the moment
this would complicate the Identity Resolution and therefore we postpone
it until later.

For #687.