apis: update cursor logic to use the most recent change time
Previously, during import, the last change time in the cursor was taken
from the last user read. However, this assumed that users were always
returned by the app in ascending order of their last change time.

Now, the last change time in the cursor is taken from the user with the
most recent last change time.