chichi: remove the `Cursor.ID` field
This change removes the 'Cursor.ID' field. The ID field was supposed to
be useful alongside the LastChangeTime field for an app to read records
for a new import, picking up from where the previous one ended. However,
it's unlikely that an app could use the last read ID to determine where
to continue reading.