chichi: state that `Cursor.LastChangeTime` is effectively in UTC
Previously, `Cursor.LastChangeTime` was assumed to be in the 'Local'
timezone. However, recent code changes have effectively set it to UTC.
This is necessary as the 'connectors' package returns the user's last
change time in UTC, making it impossible to storing it in the original
local location. This change ensures that the documentation accurately
reflects the current usage of `Cursor.LastChangeTime` as UTC.