apis/datastore: don't pass meta properties to alter schema methods
This commit documents and makes consistent the use of schema-altering
warehouse methods.

In particular, it ensures that meta properties do not have to be passed
to such methods, since this is information that each driver can handle
itself (a driver may in fact have other meta properties in addition to
those required by Chichi).

Furthermore, this simplifies the driver code, which can then
add meta properties without having to remove anything first.

This commit is in preparation for the PR #801 and is related to the
issue #795.