apis/datastore: don't pass meta properties to `RunIdentityResolution`
Currently, it is unclear whether the `RunIdentityResolution` method
should or should not take meta properties in its "userColumns"
parameter. Since it is not documented, the PostgreSQL implementation of
the method must purposely ignore the "__id__" meta property.

This commit specifies and documents the method, passing it the correct
parameters and simplifying the driver.

This commit is in preparation for the PR #801 and is related to the
issue #795.