apis/transformers/mappings: convert non-composite types to Array
This change allows for the conversion of non-composite types to the
Array type. More precisely, if T1 is a non-composite type, it can be
converted to the type Array(T2) if T1 is convertible to T2. In such
cases, the value 'v' is converted into '[]any{v}'.

For JSON values, only JSON String, Bool, and Number values undergo
conversion.

Close #793