apis/datastore: implement ack function in `IdentityWriter`
This change implements the ack function for the identity writer. Now, an
ack function is required to create an 'IdentityWriter' instance.
Additionally, the 'Write' method now accepts an ack ID (which can be
empty) and calls the ack function for individual ack IDs or batches when
a property validation fails or when the identities are written into the
data warehouse.