apis/connectors: fix property validation in the `RecordString` method
Previously, the 'RecordString' method of '*recordWriter' validated
property values using a simplified version of the 'normalize' function
tailored for the 'Text' type. However, a connector isn't obliged to call
the 'RecordString' method solely because all properties have the 'Text'
type. Instead, it's preferable to use 'RecordString' due to its
convenience, irrespective of property types.

This change addresses the issue by removing the specialized
'validateStringProperty' method and using the 'normalize' method.