apis/connectors: validate column names returned by `Columns` and `Query`
This change validates the column names returned by the Columns and Query
methods of the Database interface in connectors. If a column name is
invalid, it returns a clear error message.

Additionally, this change removes the now redundant check on column
names performed by the Columns method in the PostgreSQL connector.