apis: instantiate the `IdentityWriter` with a `*state.Action`
This commit introduces a logical change such that a list of parameters
is no longer passed to the IdentityWriter, but the Action is passed
directly (as it is in the state).

This, in addition to the implementation changes that it opens (and that
will be made in subsequent commits), clarifies at a logical level the
fact that an IdentityWriter instance refers to an action.