apis/datastore: fix initialization of stores
Currently, stores are initialized in a goroutine. This means that it is
possible that when the 'New' method (which returns a '*Datastore')
returns, some stores may not be initialized yet.

This commit causes stores to be initialized in the same goroutine as the
caller, to avoid this issue. Maybe in the future we will review the
loading of Chichi in general to optimize the timings.

Co-authored-by: Marco Gazerro <gazerro@open2b.com>