javascript-sdk: rename 'data' variables to 'event' to clarify behavior
The variable named 'data', utilized across my function, represents the
event to be sent. Therefore, this commit renames these variables to
'event' for clarity.

Additionally, the '#serveEvent' method no longer needs to return the
dispatched event since it is the same as the one passed as an argument
to the method.