apis: remove listeners when scheduler is shut down or closed
Previously, when the scheduler was shut down, the listeners were not
removed. This caused a problem where if a node was no longer the leader,
the scheduler would shut down, but the listeners it had added were not
removed. Consequently, the scheduler was not released by the garbage
collector, and notifications continued to be sent to the listeners.

This commit ensures that listeners are removed when the scheduler is
shut down or closed.