apis/state: ensure no state updates after close
This change ensures that once the '(*State).Close' method returns, no
further updates to the state will be made and no listeners will be
called. The fact remains that when the Close method is called, no calls
to the State methods should be in progress, and no more should be made.