apis/state: implement `RemoveListeners` method and refactor listeners
This change implements the RemoveListeners method to remove previously
added listeners. It also refactors listener handling by using a unique
map of listeners instead of a slice for each notification type.

It also removes the Keep method because New returns a frozen state. To
start keeping the state updated, the Unfreeze method can be called.

This is a reimplementation of commit 58cc57f, which was reverted by
commit 60fd265. This new implementation does not require locking.