apis/transformers/mappings: replace the `when` function with `if`
This change replaces the 'when' function with the 'if' function. The
'if' function takes either two or three arguments.

The first argument must be of Boolean type. If the first argument is
true, it returns the second argument. If the first argument is false, it
returns the third argument or void if there are only two arguments.