apis: fix user import from file after a user cursor reset
Previously, after resetting the cursor (e.g., due to a change in the
"identity" property), performing an import without requesting a reimport
would still read all users from the file. However, the purge was not
executed because it relied on the 'Execution.Reimport' field.

Consequently, it was expected that the purge of identities no longer
present in the file would occur even without a reimport, which was not
the case.

Fixes #876