apis/datastore: replace uses of `state.Filter` with a new `Where` type
This change updates the type of the 'Query.Where' field from
'*state.Filter' to '*Where'. Currently, the 'Where' type is identical to
the 'state.Filter' type.

The filters in actions require refactoring, and we aim to decouple them
from their usage in the datastore package. Therefore, this commit
removes this dependency.

Refers #766