doc: document how the matching in app export occurs
This commit documents how the matching in app export occurs.

This is not to say that this is the best behavior: it is simply that the
commit is intended to make the currently missing documentation
consistent with the implementation of export to apps.

It is possible that this behavior will be discussed and eventually
revised in the future.