apis: return unused properties error even for transformation functions
Previously, we didn't have information about the properties of the
input/output schema used by transformation functions, so checking the
properties used didn't make sense for actions with a transformation
function.

With the current implementation, however, we have the list of used
properties available inside the action. So this commit adds this check
for actions with transformation functions as well.

Resolve #893.