all: hide support for groups
This commit hides the support for groups, since it is not fully
implemented yet, as discussed in #892.

In general, the code that supports groups has been kept, and has simply
been made inaccessible. However, the group tables of the various
warehouse drivers have been removed, as they were obsolete and
inconsistent.

References to the issue #895 has been inserted in some parts of the code
to make easier re-enabling the groups when we will implement them.

Also see #896.

Resolve #892.