apis: fix validation of filters in method `Workspace.Users`
At the moment, the 'Workspace.Users' method accepts meta properties
within filters.

For now, to make the implementation consistent, since 'Workspace.Users'
operates on properties of the user schema, and the meta properties of
that schema must be hidden from the outside, the filter validation must
be consistent and prevent meta properties from being accepted in
filters.

When we revisit the filter implementation (#766), we may reconsider
this.