assets: fix the samples shown in the transformation testing mode
Currently, when the testing mode of the transformation is activated and the
samples represent users, the list of samples is displayed with the string
'undefined' instead of the main information that identifies each sample.

This commit solves the problem by ensuring that user properties are accessed
correctly when the samples are constructed.

Close #898.