apis/datastore: fix data warehouse mode check in event identity writer
Previously, calling (*Store).EventIdentityWriter would fail with an
error if the data warehouse was in Inspection mode or Maintenance mode.

This behavior was incorrect because the data warehouse can transition
into Inspection or Maintenance mode at any point during event reception.

Therefore, this change ensures that the data warehouse mode is only
checked when the Write method of *EventIdentityWriter is called.