assets: add string validation in `removeQuotes`
This commit adds string validation to the `removeQuotes` function, used on the
action page to remove any double quotes from the values of the sample's
properties used to show the samples in the testing mode.

This modification is necessary because it is not guaranteed that such values
will always be present, since the users (all or just some) might have been
imported without any mapping on those properties.