apis/datastore: rewrite and merge warehouses documentation in one table
This commit rewrites and updates the warehouse documentation, and merges
it into a single table, so that it is clear what is different from one
warehouse to another, what is supported and what is not.

The documentation has been rewritten and verified against the current
warehouse documentation; nevertheless, there are still some unclear
points that need to be discussed. So, as also stated in the Markdown
file, for now the table is not reliable and needs to be further
discussed and revised.