apis/datastore: update the types support in `README.md`
This commit modifies the datastore `README.md` file by revising the
principles by which we support certain types and the types themselves;
the table should therefore be quite correct, while the implementation of
both the drivers, the checks on the user schema, and the data warehouse
drivers, still needs to be revised.

For #922.