all: when exporting to file, put users schema in input schema
This commit changes the structure of actions that export users to file
so that the user schema is in the input schema and not in the output
schema, as it is now.

This change seems to be logical and consistent with the new
implementation of validation and alignment checks. It also avoids having
to handle the export to file as a special case.