apis: add `PreserveJSON` field to choose JSON value preservation
This change introduces a PreserveJSON field to the transformation
function. This field determines whether JSON values should be preserved
in their original format.
      
When PreserveJSON is set to true, JSON values are passed to the
transformation function as strings representing the JSON source. The
function must return JSON as strings.
      
When PreserveJSON is set to false, JSON values are passed to the
function as their corresponding JavaScript or Python values (e.g., null,
false, true, string, number, object, and array for JavaScript; None,
False, True, str, int, float, dict, and list for Python). The
transformation function must return values in the same format. 

Specifically, nil and JSON null values are passed as JavaScript null or
Python None. If the function returns JavaScript null or Python None, it
will be interpreted as JSON null when PreserveJSON is false.

Closes #877

Co-authored-by: Andrea Retini <retini@open2b.com>