assets: remove save button handling from connector UI
Previously, when the UI of a connector was displayed, if there was a
button with a "save" event, it would be shown. Otherwise, a specific
alternative button would be displayed.

Since it has been decided not to allow connectors to define a "save"
button (see commit 8021082) and since no connector currently defines a
"save" button (see commit 91c9db1), this change removes the code that
handles the "save" button from the UI of the connectors.