all: disallow leading and trailing spaces in `json.RawMessage` values
In the Meergo source code, it is common practice to check the first
character of a 'json.RawMessage' as a quick and effective method to
determine the type of JSON value it represents. For example, checking if
value[0] == '{' tests whether the value is a JSON Object. However, if
the value contains leading spaces, this validation would fail, leading
to unexpected consequences.

To address this, this change disallows both leading and trailing spaces
in 'json.RawMessage' values when used in JSON property values.