assets: handle external matching property in transformation functions
Currently, when using the transformation function, all possible transformable
properties are automatically assigned to the action's output schema.

However, since external matching properties cannot be used in transformations,
this behavior leads to errors that make it impossible to save the actions.

This commit fixes the issue by preventing the external matching property from
automatically ending up in the output schema.