apis/events/collector: avoid setting `Context.Traits` to an empty map
Previously, the 'Context.Traits' field of an event was set to an empty
map if it was nil. This change leaves the value as nil, but continues to
store a JSON object in the data warehouse.

This change is necessary to distinguish whether 'Context.Traits' was
passed with the event or not.