apis/events/collector: add `collectedEvent.originalTimeStamp` field
This change introduces the 'collectedEvent.originalTimeStamp' field,
which captures the time the event was sent, based on the timestamp
provided by the client. While the timestamp field is adjusted for skew
(receivedAt - sendAt), the originalTimeStamp preserves the original
timestamp as received from the client.

This field should be serialized in JSON in the
'(*Observer).addCollectedEvent' method but currently is not.