apis/events: refactor event schema properties
This commit refactors the event schema by making some properties read
optional and changing the nullability of others.

The properties that are not read optional, and are therefore always
available in a transformation, are:

* anonymousId
* context
* messageId
* receivedAt
* sentAt
* source
* timestamp
* type

Also, not read optional are the second-level properties under 'context',
such as 'context.app.name' if 'app' is present. As an exception,
'context.session.start' is read optional.

The 'userId' property is now the only nullable property. Previously,
'properties' and 'traits' were also nullable.

Closes #932