apis: fix external matching property value when exporting to apps
The external matching property is not included in the output schema of a
transformation; instead, it is directly copied from the internal
matching property with its external name and type.

However, the export was incorrect because the external matching property
value was set before transforming the other properties, rather than
after. As a result, the transformation could receive the external name
and type instead of the internal one. If the transformation succeeded,
the external matching property would still not be included among the
transformed properties and, therefore, would not be passed to the app
during creation.