apis/datastore/warehouses/postgresql: clarify "users" view creation
This commit ensures that the REPLACE of the view only happens when
necessary, specifically during Identity Resolution, and not when the
schema of the users table is altered. In the latter case, it is
necessary to discard the view and rebuild it from scratch, as the table
needs to be altered and it would be impossible to alter a column if
there is a view that references it.

It also improves the related documentation and updates the tests
accordingly.