connectors/dummy: reduce timestamps precision to microseconds
As indicated in commit eb6c635d4af13729a6351bbf3a5e8992b024bacd, there
is currently a discrepancy in the precision of the cursors managed by
Meergo, based on the timestamps returned by the connectors, and those
stored in the PostgreSQL database. This commit aims to address this
discrepancy by ensuring that Dummy returns timestamps with microsecond
precision instead of nanosecond precision, as is currently the case.