apis/datastore/warehouses/postgresql: fix panic
This commit prevents a panic that occurs because the PostgreSQL driver
incorrectly passes 'warehouse.db' as a parameter to the
'identityResolutionExecution' function, which may not be initialized.

This is noticeable when 'IdentityResolutionExecution' is called as the
first method, as it tries to access a field that is still nil, causing a
panic.

This commit fixes the code for the call.