apis/datastore: limit wh. operations to the ones allowed in current mode
Add a "mode coordinator" to the datastore so that only operations
compatible with the current mode can be initiated, and ensures that
during mode switch, operations incompatible with the new mode are
cancelled.

It should be noted that, after various discussions, we have established
that the concept of 'warehouse mode' probably needs to be reviewed. In
this regard, an issue has been opened, #1002, so the changes introduced
by this commit, although they should be correct and consistent with each
other, may be reviewed later.