all: refactor the data warehouse connection
This commit refactors the way that data warehouses are connected.

In particular, it makes the "connect warehouse" method handle with
three different behaviors:

* FailOnCheck, to make the warehouse connection check if the warehouse
  objects are present; in this case, initialization or repairing errors
  may be returned.
* InitializeWarehouse, to make the warehouse connection initialize the
  warehouse before connecting it;
* RepairWarehouse, to make the warehouse connection repair the warehouse
  before connecting it.

As a consequence, this commit eliminates the method "init warehouse".

This commit also changes the APIs at any level, updates the
documentation and the "meergo-cli" tool to support this new behavior.

Only the "FailOnCheck" behavior is supported by the UI currently. This
is discussed in the issue #1007.

Also take a look at the issue #1008.