assets: restrict testing mode for unmatched properties
This commit ensures that when an action requires matching properties, access to
the testing mode will be blocked until these properties are defined. Attempting
to open testing mode without them will result in an error.

This change is necessary to prevent errors related to matching properties during
the testing phase, as such errors would be less clear in that context.

Close #1021.