apis/state: fix addition of workspace
When a new workspace is added to the state, its 'UserPrimarySources' and
'Identifiers' fields are not initialized and remain 'nil', causing an
inconsistency in the state that is then exposed through the APIs and
leads to errors in the UI as well.

This commit fixes the handling of the 'AddWorkspace' notification to
resolve the issue.