all: remove the default workspace
This commit removes the default workspace that is currently added by the
PostgreSQL.sql script, which generates the database. Consequently, it
modifies everything necessary to allow Meergo to start and create a
workspace.

Note that the UI currently does not handle the deletion and recreation
of the workspace correctly. See issue #1025.

Moreover, the issue #1026 discusses rearranging the meergo-cli commands.