connectors/mysql: add a test for the `Upsert` and `Query` methods
This change adds a test to validate all supported MySQL values when
using the Upsert and Query methods and updates the propertyType
function.

It does not run automatically with the Meergo tests but requires the
MEERGO_TEST_PATH_MYSQL environment variable to be set.

Additionally, it updates the PostgreSQL and ClickHouse tests.