test: improve test on Identity Resolution
This commit improves the 'TestIdentityResolution2' test by avoiding the
automatic execution of Identity Resolution, instead starting the process
only explicitly.

This ensures that multiple Identity Resolutions are not started
concurrently, a situation in which those started later should return an
error, which is something not covered nor handled by this test.

So, this commit is in preparation for the fixes on the execution of
operations on the data warehouse.