assets: fix error when opening the Analytics tab with event actions only
This commit fixes an issue that occurred when opening the "Analytics" tab of a
connection without any actions related to users.

Since by default statistics were requested for users, even when there were no
actions for them, the statistics request would return an error, leading to
unexpected behavior in the UI.

This commit fixes the issue by ensuring that when there are no actions related
to users, only event statistics are requested.

Close #1043.