all: expose the Repair warehouse operation through a dedicated method
In the past, the Repair operation was implemented to work in an
integrated manner with the connection of a warehouse; then the warehouse
connection was rewritten, and therefore the repair was removed as an
operation. Finally, with the discussion that led to issue #1017, it was
decided to make the repair operation available through a dedicated
method via the APIs.

This commit therefore exposes the Repair through the APIs, modifies and
tests, and modifies 'meergo-cli' so that it is possible to initiate this
operation from the command line as well.

The implementation of the possibility to initiate the Repair from the UI
will be discussed separately.

For #1017.