apis,json: refactor `json.Value.Lookup` and remove `gjson` dependency
This change refactors the Value.Lookup method, removing the dependency
on gjson and accepting the path as a []string. Additionally, it
refactors the code that uses this method.

The type of the Property field in the state.Where type is changed from
string to []string, allowing it to be passed to the Lookup function
without conversion to a string.

The part type in mapping has been refactored to separate the path from
its decorators, such as "x?" and "[x]", so that the path without
decorators can be passed directly to the Lookup method.

The Lookup method has been designed to differentiate, based on the
returned error, whether an attempt was made to access a non-object value
by key or if the key simply does not exist in the object.

The Value.Get method has been added to retrieve a value based on a path,
returning a boolean instead of an error.