warehouses/postgresql: simplify and optimize IR removing "same_user"
This commit simplifies and optimizes identity resolution in the
PostgreSQL driver by removing the SQL function "same_user" and instead
handling that logic directly in Go.

This not only simplifies the implementation of Identity Resolution but
also results in optimization as it avoids the definition and calling of
a function for each matching.

Although it is difficult to quantify the actual performance improvement
since these tests were conducted on a desktop PC with various
applications running, here are the execution times of Identity
Resolution on a dataset of 5000 users, with three identifiers
configured.

A reduction in execution times of approximately 23% has been achieved.

In particular:

Times without modification (on main):

    00:01:41.881203
    00:02:05.44844

    Avg: ~ 111 seconds.

Times with the modification from this commit:

    00:00:50.425845
    00:01:41.111513
    00:01:30.740798
    00:01:42.165245

    Avg: ~86 seconds.