warehouses/postgresql: simplify and drastically optimize Identity Resol.
This commit simplifies the 'edges' table by removing a boolean column,
which consequently means that it is no longer necessary to store
unnecessary rows. The logical behavior of the Identity Resolution (IR)
has not been altered.

This results in, in addition to the simplification, a drastic reduction
in space occupied and execution times of the Identity Resolution.

In particular, in a test conducted locally on my PC with 5,000 user
identities imported from CSV, the following results were obtained:

Before this change:

* The 'edges' table had 12,497,500 rows.
* The execution time of the IR was about 2 minutes.

After this change:

* The 'edges' table has 18,841 rows.
* The execution time is about 2 seconds.

Thus, the number of rows in the 'edges' table is now 0.15% of what it
was before (a reduction by a factor of one thousand), while the
execution speed is 60 times faster compared to before the change.