core/datastore: add test on Identity Resolution behavior on warehouses
This commit adds a test that verifies the correct behavior of Identity
Resolution through calls to the methods of the warehouse drivers.

To run the test, please refer to the documentation in the file:

    core/datastore/warehouses_identity_resolution_test.go

A unified framework for executing this type of test will likely be
discussed and introduced later.

Note that the purpose of this commit is not to provide a complete set of
tests covering all behaviors of Identity Resolution. Additional tests
will be added in subsequent commits.