assets: review the style of the users grid
This commit fixes some issues related to the users grid styling.

Specifically, it fixes the vertical overflow of the grid on the
viewport, it standardize the radius of the grid borders, and it prevents
the grid footer from overlapping the overlay that opens when clicking on
the "Resolve identities" button.

Close #1091.