warehouses/snowflake: quote table and column names in table definitions
The idea is to ensure that all identifiers used in Snowflake (table
names, column names, etc.) are always double quoted.

This is necessary to avoid errors, as in Snowflake the double quoted
identifier:

     "hello"

is different from the identifier:

     hello

Subsequent commits could correct other parts of the driver.

For more details:

     https://docs.snowflake.com/en/sql-reference/identifiers-syntax#double-quoted-identifiers