warehouses/snowflake: review quoting of identifiers
The purpose of this commit is to roughly correct the quoting in order to
avoid various scattered errors in the driver when trying to access a
table or a column.

In the future, subsequent commits may review these changes in details,
comparing with and reviewing the PostgreSQL driver as well.