warehouses/snowflake: implement a draft of the alter schema methods
This commit implements a draft of the alter schema on Snowflake. The aim
is to add the entire infrastructure of methods that allow for simple
additions of properties to the user schema (e.g., properties of type
Text).

A review of the functionality and the use of the correct types will be
carried out in subsequent commits.